MAJAA code for Assignment 4
This is a MAJAA code for assignment 4 that does the following sambavams in a sirappana way:
1). Support for key 1 of 4x4 keypad to toggle between start and stop commands
2). Support for key 2 of 4x4 keypad to toggle between resume and pause commands
3). Support for Seven Segment LED displays

What the code needs to do (OPTIONAL):
Since delay has been added for Seven segment display, we can try reducing the blink delays for more accurate operation. But I believe this is just optional.