Updated Assignment 6 code
Everything works as per the requirement of the assignment. But when I tried to to configure interrupts for PORTC, the program crashed. So for now, I have added the tic tac toe in systick handler function. It still works. Try enabling interrupts for PORT C and try to activate the tic tac toe game through the interrupts for PORTC. For now, this is OK. But do try to implement interrupts for PORTC.